REST: route collection mutations through DocumentsController
