+++ docs/tech-debt.md
@@ -495,7 +495,7 @@ That is acceptable for the current DataView scale. It is not a real bulk operati
 
 Drag/drop and `menu_order` accounting look at both pages and collections through `treeRecords`, a single `DndContext` now wraps both sections, and the cycle guard walks the merged node graph. The model gap that remains is shape, not wiring: every tree consumer still branches on pages versus collections to derive a `kind`, and the Collections section's contents are a UX choice on top of that. Today it shows only top-level full-page collections (`parent = 0`); nested ones live solely in the Pages tree. That follows Notion's "one thing, one place" rule, but once collections nest there is no surface for "see every collection at a glance". The unified model would let this be configurable per workspace: top-level only, all collections grouped, or per-parent sub-headers. Lazy loading lives in RSM-1848.
 
-**Where.** `ACTIVE_PAGES_QUERY` in `src/components/page-queries.js`, `FULL_PAGE_COLLECTION_QUERY` in `src/collections.js`, `nestedCollections` / `topLevelCollections`, `treeRecords`, `handleDragOver`, and the shared `DndContext` in `src/components/Sidebar.js`, the `renderCollectionRow` bridge in `src/components/PageRow.js`, `CollectionRow`'s drag/drop zones, `computeDropTarget` in `src/components/pages-tree.js`, and the `Collection::hierarchical` / `CollectionsController::validate_parent_document` setup on the PHP side.
+**Where.** `ACTIVE_PAGES_QUERY` in `src/components/page-queries.js`, `FULL_PAGE_COLLECTION_QUERY` in `src/collections.js`, `nestedCollections` / `topLevelCollections`, `treeRecords`, `handleDragOver`, and the shared `DndContext` in `src/components/Sidebar.js`, the `renderCollectionRow` bridge in `src/components/PageRow.js`, `CollectionRow`'s drag/drop zones, `computeDropTarget` in `src/components/pages-tree.js`, and the `Collection::hierarchical` / `Collection::validate_parent_document` setup on the PHP side.
 
 **Solution.** Add a workspace-tree REST endpoint that returns navigable nodes with `kind`, `id`, `parent`, `menu_order`, and visibility in one shape. Row-owned collections could then appear under their row, missing parents would not look top-level by accident, and page/collection branching would move out of every consumer. The Collections section then becomes a view over the same model with the user's chosen filter, instead of a separate flat list.
 
@@ -503,7 +503,7 @@ Drag/drop and `menu_order` accounting look at both pages and collections through
 
 **What.** Duplicating a full-page collection creates the new collection, registers its row CPT, and copies field posts that stand on their own. It skips relations because a relation is really a pair: the forward field plus the reverse field on another collection. A safe copy has to create or update both sides, keep the cardinality, and remap ids without touching the original relation. Until that exists, the REST response lists skipped fields and the sidebar tells the user the copy is missing columns. Rollups that read through a skipped relation belong in the same bucket; they are not useful until the copied schema has its own relation target.
 
-**Where.** `CollectionsController::duplicate()`, `clone_fields()`, and `remap_rollup_references()` in `includes/Rest/CollectionsController.php`, plus the skipped-field notice in `src/components/Sidebar.js` and duplicate coverage in `tests/php/test-rest-collections-controller.php`.
+**Where.** `CollectionDuplicator::duplicate()`, `clone_fields()`, and `remap_rollup_references()` in `includes/Documents/CollectionDuplicator.php`, plus the skipped-field notice in `src/components/Sidebar.js` and duplicate coverage in `tests/php/test-rest-collections.php`.
 
 **Solution.** Add a relation-aware schema copy step. It should clone and remap the forward and reverse fields together, or skip every dependent field, including rollups that point at skipped relations. The duplicate should never carry references back to the source collection's fields. Once that exists, the sidebar notice can name the exact skipped field types instead of treating them all as generic missing columns.
