+++ includes/Rest/DocumentsController.php
@@ -2,13 +2,12 @@
 /**
  * REST endpoints for Cortext documents.
  *
- * Lists every post type that opts into the `cortext-document` trait (pages
- * plus collection rows) from one endpoint, so callers do not need to rebuild
- * the post-type list. Also owns restore and permanent-delete for documents in
- * trash; both ask `TrashCascadeEngine` for the page hierarchy snapshot so
- * descendants move with their root.
+ * Handles listing, restore, permanent delete, duplicate, and collection
+ * create through the `Documents` service.
  *
- * Single-resource reads still go through `DocumentLocatorController` plus
+ * Restore and permanent-delete walk the page hierarchy via
+ * `TrashCascadeEngine` so descendants move with their root. Single-resource
+ * reads still go through `DocumentLocatorController` plus
  * `/wp/v2/<rest_base>/<id>`.
  *
  * @package Cortext
@@ -22,8 +21,8 @@ use Cortext\Documents;
 use Cortext\PostType\Cascade\CollectionToRowTrashCascade;
 use Cortext\PostType\Cascade\DocumentToCollectionTrashCascade;
 use Cortext\PostType\Cascade\PageHierarchyTrashCascade;
+use Cortext\PostType\Collection;
 use Cortext\PostType\CollectionEntries;
-use Cortext\PostType\Page;
 use Cortext\PostType\TrashCascadeEngine;
 use WP_Error;
 use WP_Post;
@@ -121,6 +120,49 @@ final class DocumentsController {
 					'callback'            => array( $this, 'permanent_delete' ),
 					'permission_callback' => array( $this, 'check_document_post' ),
 					'args'                => $id_arg,
+				),
+			)
+		);
+
+		register_rest_route(
+			self::NAMESPACE,
+			'/documents/(?P<id>\d+)/duplicate',
+			array(
+				array(
+					'methods'             => 'POST',
+					'callback'            => array( $this, 'duplicate' ),
+					'permission_callback' => array( $this, 'can_duplicate' ),
+					'args'                => $id_arg,
+				),
+			)
+		);
+
+		// Collection create keeps the kind in the URL instead of the request
+		// body. Page create still uses `/wp/v2` through core-data.
+		register_rest_route(
+			self::NAMESPACE,
+			'/collections',
+			array(
+				array(
+					'methods'             => 'POST',
+					'callback'            => array( $this, 'create_collection' ),
+					'permission_callback' => array( $this, 'can_create' ),
+					'args'                => array(
+						'title'  => array(
+							'type'     => 'string',
+							'required' => true,
+						),
+						'mode'   => array(
+							'type'    => 'string',
+							'enum'    => array( Collection::MODE_INLINE, Collection::MODE_FULL_PAGE ),
+							'default' => Collection::MODE_FULL_PAGE,
+						),
+						'parent' => array(
+							'type'    => 'integer',
+							'minimum' => 0,
+							'default' => 0,
+						),
+					),
 				),
 			)
 		);
@@ -272,6 +314,58 @@ final class DocumentsController {
 			),
 			200
 		);
+	}
+
+	public function can_create(): bool {
+		return current_user_can( 'edit_posts' );
+	}
+
+	/**
+	 * Permission gate for duplicate. Mirrors `check_document_post` so a
+	 * missing document returns 404 before capability checks can turn it into
+	 * a generic forbidden response.
+	 *
+	 * @param WP_REST_Request $request Incoming REST request.
+	 *
+	 * @return bool|WP_Error
+	 */
+	public function can_duplicate( WP_REST_Request $request ) {
+		$id   = (int) $request->get_param( 'id' );
+		$post = get_post( $id );
+
+		if ( ! $post instanceof WP_Post || ! post_type_supports( $post->post_type, 'cortext-document' ) ) {
+			return new WP_Error(
+				'cortext_document_not_found',
+				__( 'Document not found.', 'cortext' ),
+				array( 'status' => 404 )
+			);
+		}
+
+		return current_user_can( 'edit_posts' ) && current_user_can( 'edit_post', $id );
+	}
+
+	public function create_collection( WP_REST_Request $request ): WP_REST_Response|WP_Error {
+		$result = $this->documents->create_collection(
+			(string) $request->get_param( 'title' ),
+			(string) $request->get_param( 'mode' ),
+			(int) $request->get_param( 'parent' )
+		);
+
+		if ( $result instanceof WP_Error ) {
+			return $result;
+		}
+
+		return new WP_REST_Response( $result, 201 );
+	}
+
+	public function duplicate( WP_REST_Request $request ): WP_REST_Response|WP_Error {
+		$result = $this->documents->duplicate( (int) $request->get_param( 'id' ) );
+
+		if ( $result instanceof WP_Error ) {
+			return $result;
+		}
+
+		return new WP_REST_Response( $result, 201 );
 	}
 
 	/**
